requirements.txt
faker
google-adk
gradio
google-cloud-bigquery
google-cloud-aiplatform
python-dotenv
